Extract routePathNeighborLinkService to new file
